LayoutTests/imported/w3c/web-platform-tests/trusted-types/TrustedTypePolicyFactory-getAttributeType-namespace-expected.txt


FAIL 0: getAttributeType with full namespace info. assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 0: getAttributeType with element namespace and empty attribute namespace assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 0: getAttributeType without namespaces. assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 0: getAttributeType with undefined and empty namespace. assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 0: getAttributeType with empty and undefined namespace. assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 0: getAttributeType with empty namespaces. assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 0: getAttributeType with element namespace and empty attribute namespace. assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 1: getAttributeType with full namespace info. assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 1: getAttributeType with element namespace and empty attribute namespace assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 1: getAttributeType without namespaces. assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 1: getAttributeType with undefined and empty namespace. assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 1: getAttributeType with empty and undefined namespace. assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 1: getAttributeType with empty namespaces. assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 1: getAttributeType with element namespace and empty attribute namespace. assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 2: getAttributeType with full namespace info. assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 2: getAttributeType with element namespace and empty attribute namespace assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 2: getAttributeType without namespaces. assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 2: getAttributeType with undefined and empty namespace. assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 2: getAttributeType with empty and undefined namespace. assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 2: getAttributeType with empty namespaces. assert_equals: expected (string) "TrustedScriptURL" but got (object) null
FAIL 2: getAttributeType with element namespace and empty attribute namespace. assert_equals: expected (string) "TrustedScriptURL" but got (object) null
